filebuf: make unlocking atomic
When renaming a lock file to its final location, we need to make sure
that it is replaced atomically. There are situations where this is not
possible in all OSs, so we try to make the best of it by removing the
target file in cases where we do not have a lock.